// src/app/core/enums/offerstatus.enum.ts
export enum OfferStatusEnum {
    zero = 'All',
    ten = 'In Progress',
    eleven = 'Waiting for Review (BD)',
    threen = 'Waiting for Cost (PE) ',
    fifteen = 'Waiting for Quote (BD)',
    seventeen = 'Quote Generated',
    // fifteen = 'In Progress',
    twenty = 'Sent For Approval',
    thirty = 'Rejected',
    fourty = 'Approved',
    fifty = 'Offer Generated',
    fiftyfour = 'Waiting for MFG/TRD Articles',
    fiftyfive = 'Waiting for SPR Articles',
    fiftysix = 'CPO Validation',
    fiftyseven = 'Sent For FM Approval',
    fiftynine = 'Rejected',
    sixty = 'Ready to Initiate SO',
    sixtytwo = 'CPO Validation',
    seventy = 'SO Initiated',
    seventytwo = 'SO Initiated',
    seventyone = 'Sent For Customer Approval',
    seventythree = 'Amendment Request',
    seventyfour = 'SAP In-progress',
    seventyfive = 'SAP Error(s)',
    eighty = 'SO Created'
}

// 56: offer CPO validation
// 59: fm rejected
// 62: CPO order CPO validation